move app and refresher to separate files
